Updated to the last engine version and fixed collisions and visibility

+++ src/playState.js
@@ -7,6 +7,7 @@ function playState() {
     var that = Object.create(FM.state()),
         avatarType = FM.objectType("avatar"),
         wallType = FM.objectType("wall"),
+        visibilityObjectType = FM.objectType("visibilityObject"),
         player = avatar(avatarType),
         eye = FM.gameObject(98),
         iris = FM.gameObject(99),
@@ -22,6 +23,7 @@ function playState() {
         cursorRenderer,
         cursorDistance,
         eyeSpatial,
+        eyeRenderer,
         eyePhysic,
         irisSpatial,
         irisRenderer,
@@ -34,13 +36,13 @@ function playState() {
         maxWalls = 2,
         move = false,
         dead = false,
-        invisible = false,
+        invisible = true,
         checkingVisibility = false,
         finalGoalReached = false;
 
     /**
-    * Initialize the play state
-    */
+     * Initialize the play state
+     */
     that.init = function () {
         //Setting the bounds of the world
         Object.getPrototypeOf(that).init(1024, 768);
@@ -56,30 +58,41 @@ function playState() {
         //ambianceAudio.play("ambiance", 0.2, true);
 
         //Loading tmx file
-        var map = FM.tmxMap(), tilemap, irisDist, i, body, wall;
-        map.load(FM.assetManager.getAssetByName("world").getContent());
+        var background,
+            irisDist,
+            renderer,
+            i,
+            body,
+            wall;
+        //map.load(FM.assetManager.getAssetByName("world").getContent());
         //Load tiles
-        tilemap = FM.tileMap(FM.assetManager.getAssetByName("ground"), 38, 38, 32, 32, [], 1, false);
-        that.getWorld().loadTileMap(tilemap, map, "ground", "ground");
+        //tilemap = FM.tileMap(FM.assetManager.getAssetByName("ground"), 38, 38, 32, 32, [], 1, false);
+        //that.getWorld().loadTileMap(tilemap, map, "ground", "ground");
+        background = FM.gameObject(1);
+        FM.spatialComponent(0, 0, background);
+        FM.spriteRendererComponent(FM.assetManager.getAssetByName("background"), 1024, 768, background);
         //Cursor
         cursorSpatial = FM.spatialComponent(FM.game.getMouseX(), FM.game.getMouseY(), cursor);
         cursorRenderer = FM.animatedSpriteRendererComponent(FM.assetManager.getAssetByName("cursor"), 11, 21, cursor);
         cursorRenderer.addAnimation("default", [0, 1, 2, 3, 2, 1, 0], 20, true);
         cursorRenderer.play("default");
         //Eye
-        eyeSpatial = FM.spatialComponent(474.5, 346.5, eye);
-        FM.spriteRendererComponent(FM.assetManager.getAssetByName("eye"), 75, 75, eye);
+        eyeSpatial = FM.spatialComponent(474, 346, eye);
+        eyeRenderer = FM.spriteRendererComponent(FM.assetManager.getAssetByName("eye"), 300, 300, eye);
+        eyeRenderer.changeSize(0.25);
         eyePhysic = FM.circleComponent(18, eye);
-		eyePhysic.offset.x = 37.5 - 18;
-		eyePhysic.offset.y = 37.5 - 18;
+        eyePhysic.offset.x = 37.5 - 18;
+        eyePhysic.offset.y = 37.5 - 18;
         //Iris
         irisSpatial = FM.spatialComponent(eyeSpatial.position.x + eyePhysic.radius, eyeSpatial.position.y + eyePhysic.radius, iris);
-        irisRenderer = FM.spriteRendererComponent(FM.assetManager.getAssetByName("iris"), 30, 30, iris);
+        irisRenderer = FM.spriteRendererComponent(FM.assetManager.getAssetByName("iris"), 120, 120, iris);
+        irisRenderer.changeSize(0.25);
         irisPhysic = FM.circleComponent(15, iris);
         irisPath = FM.simplePathComponent(iris);
         irisAudio = FM.audioComponent(iris);
         irisAudio.addSound(FM.assetManager.getAssetByName("visible"));
         //Visibility test object
+        visibilityTestObject.addType(visibilityObjectType);
         visibilityTestObjectSpatial = FM.spatialComponent(irisSpatial.position.x + irisPhysic.radius - 1, irisSpatial.position.y + irisPhysic.radius - 1, visibilityTestObject);
         visibilityTestObjectPhysic = FM.aabbComponent(2, 2, visibilityTestObject);
         visibilityTestObjectPath = FM.simplePathComponent(visibilityTestObject);
@@ -92,7 +105,7 @@ function playState() {
         }
         for (i = 0; i < persistentData.walls.length; i = i + 1) {
             wall = FM.gameObject(15);
-            console.log(persistentData.walls[i]);
+            wall.addType(wallType);
             FM.spatialComponent(Math.round(persistentData.walls[i].x), Math.round(persistentData.walls[i].y), wall);
             if (persistentData.walls[i].inclined === "true") {
                 FM.spriteRendererComponent(FM.assetManager.getAssetByName("wallInclined"), 10, 100, wall);
@@ -106,6 +119,7 @@ function playState() {
         }
 
         //Add objects
+        that.add(background);
         that.add(player);
         that.add(eye);
         that.add(iris);
@@ -115,8 +129,8 @@ function playState() {
     };
 
     /**
-    * Update the play state
-    */
+     * Update the play state
+     */
     that.update = function (dt) {
         Object.getPrototypeOf(that).update(dt);
 
@@ -125,7 +139,7 @@ function playState() {
             player.audio.play("walk", 1, false);
             move = true;
             player.path.clearPath();
-            player.path.add(FM.game.getMouseX() + player.physic.offset.x - player.physic.radius , FM.game.getMouseY() + player.physic.offset.y + player.physic.radius);
+            player.path.add(FM.game.getMouseX(), FM.game.getMouseY());
             player.path.startFollowingPath(40);
             cursorDistance = FM.vector(player.spatial.position.x - (FM.game.getMouseX() - player.physic.offset.x - player.physic.radius), player.spatial.position.y - (FM.game.getMouseY() - player.physic.offset.y));
             if (cursorDistance.y < 0) {
@@ -184,79 +198,77 @@ function playState() {
             wallsCount += 1;
         }
         //Eye movement following
-        if (!invisible) {
-            irisAudio.play("visible", 1, true);
+        if (!invisible && !finalGoalReached && !dead) {
+            if (!irisAudio.isPlaying("visible")) {
+                irisAudio.play("visible", 1, true);
+            }
             irisPath.clearPath();
             irisPath.add(player.spatial.position.x + player.physic.offset.x + player.physic.radius, player.spatial.position.y + player.physic.offset.y + player.physic.radius);
             irisPath.startFollowingPath(300);
-            if (player.spatial.position.x - irisPhysic.width < irisSpatial.position.x
-                    && irisSpatial.position.x <= eyeSpatial.position.x) {
+            if (player.spatial.position.x - irisPhysic.width < irisSpatial.position.x && irisSpatial.position.x <= eyeSpatial.position.x) {
                 irisPhysic.velocity.x = 0;
             }
-            if (player.spatial.position.x > irisSpatial.position.x
-                    && irisSpatial.position.x + irisPhysic.width >= eyeSpatial.position.x + eyePhysic.width) {
+            if (player.spatial.position.x > irisSpatial.position.x && irisSpatial.position.x + irisPhysic.width >= eyeSpatial.position.x + eyePhysic.width) {
                 irisPhysic.velocity.x = 0;
             }
-            if (player.spatial.position.y - irisPhysic.height < irisSpatial.position.y
-                    && irisSpatial.position.y <= eyeSpatial.position.y) {
+            if (player.spatial.position.y - irisPhysic.height < irisSpatial.position.y && irisSpatial.position.y <= eyeSpatial.position.y) {
                 irisPhysic.velocity.y = 0;
             }
-            if (player.spatial.position.y > irisSpatial.position.y
-                    && irisSpatial.position.y + irisPhysic.height >= eyeSpatial.position.y + eyePhysic.height) {
+            if (player.spatial.position.y > irisSpatial.position.y && irisSpatial.position.y + irisPhysic.height >= eyeSpatial.position.y + eyePhysic.height) {
                 irisPhysic.velocity.y = 0;
             }
             //Update time watched
             timeWatched += dt;
-        } else {
-            irisAudio.pause("visible");
+        } else if (!finalGoalReached && !dead) {
+            if (irisAudio.isPlaying("visible")) {
+                irisAudio.pause("visible");
+            }
             irisPath.clearPath();
             irisPath.stopFollowingPath();
-            irisSpatial.position.reset(eyeSpatial.position.x + eyePhysic.radius - irisPhysic.radius, eyeSpatial.position.y + eyePhysic.radius - irisPhysic.radius);
-            //irisPath.add(eyeSpatial.position.x + eyePhysic.radius, eyeSpatial.position.y + eyePhysic.radius);
-            //irisPath.startFollowingPath(300);
+            irisSpatial.position.reset(Math.round(eyeSpatial.position.x + eyeRenderer.getWidth() / 2 - irisPhysic.radius), Math.round(eyeSpatial.position.y + eyeRenderer.getWidth() / 2 - irisPhysic.radius));
         }
         //Reducing avatar's size
-        if (!dead && !checkingVisibility) {
+        if (!finalGoalReached && !dead && !checkingVisibility) {
             visibilityTestObjectPath.clearPath();
-			visibilityTestObjectPath.add(player.spatial.position.x + player.physic.offset.x + player.physic.radius, player.spatial.position.y + player.physic.offset.y + player.physic.radius);
-            visibilityTestObjectPath.startFollowingPath(300);
+            visibilityTestObjectPath.add(player.spatial.position.x + player.physic.offset.x + player.physic.radius, player.spatial.position.y + player.physic.offset.y + player.physic.radius);
+            visibilityTestObjectPath.startFollowingPath(400);
             checkingVisibility = true;
-        } else if (!dead && visibilityTestObjectPath.isLastWaypointReached()) {
+        } else if (!finalGoalReached && !dead && visibilityTestObjectPath.isLastWaypointReached()) {
             var cursorDistance = FM.vector((player.spatial.position.x + player.physic.offset.x + player.physic.radius) - eyeSpatial.position.x, (player.spatial.position.y + player.physic.offset.y + player.physic.radius) - eyeSpatial.position.y);
             cursorDistance = cursorDistance.getLength();
             var factor = 1 - (timeWatched / maxTimeWatched),
                 distFactor = cursorDistance / that.getWorld().width / 2;
+            //console.log(player.physic.radius);
             player.renderer.changeSize(factor);
-            player.physic.radius = 20 * factor;
-            player.physic.offset.x += 0.8;
-            player.physic.offset.y += 0.5;
+            player.physic.radius = player.physic.radius * factor;
             player.physic.width = player.physic.radius * 2;
             player.physic.height = player.physic.radius * 2;
+            player.physic.offset.x += 1.6;
+            player.physic.offset.y += 1.2;
             checkingVisibility = false;
             invisible = false;
             visibilityTestObjectSpatial.position.reset(irisSpatial.position.x + irisPhysic.radius - 1, irisSpatial.position.y + irisPhysic.radius - 1);
         }
-        var i, wall;
-        for (i = 0; i < walls.length; i = i + 1) {
-            wall = walls[i];
-            //console.log(wall.components[FM.componentTypes.SPATIAL].position);
-            if (wall.components[FM.componentTypes.PHYSIC].overlapsWithObject(visibilityTestObjectPhysic)) {
-                //console.log("qksdkjqhsdkhqsjfdgsjdq");
-                visibilityTestObjectPath.stopFollowingPath();
-                checkingVisibility = false;
-                visibilityTestObjectSpatial.position.reset(irisSpatial.position.x + irisPhysic.radius - 1, irisSpatial.position.y + irisPhysic.radius - 1);
-                invisible = true;
-            }
+        //Check collision with walls
+        if (wallType.overlapsWithObject(visibilityTestObject)) {
+            visibilityTestObjectPath.stopFollowingPath();
+            checkingVisibility = false;
+            visibilityTestObjectSpatial.position.reset(irisSpatial.position.x + irisPhysic.radius - 1, irisSpatial.position.y + irisPhysic.radius - 1);
+            invisible = true;
         }
         //Check if dead, send data and end the game
-        if (timeWatched >= maxTimeWatched && !dead) {
+        if (timeWatched >= maxTimeWatched && !dead && !finalGoalReached) {
             console.log("DEAD");
             dead = true;
             move = false;
             player.renderer.play("idle");
             player.kill();
+            player.hide();
             invisible = true;
-            irisSpatial.position.reset(eyeSpatial.position.x + eyePhysic.radius - irisPhysic.radius, eyeSpatial.position.y + eyePhysic.radius - irisPhysic.radius);
+            irisRenderer.changeSize(1.5);
+            irisSpatial.position.reset(eyeSpatial.position.x + eyePhysic.radius - irisRenderer.getWidth() / 2, eyeSpatial.position.y + eyePhysic.radius - irisRenderer.getHeight() / 2);
+            irisPath.clearPath();
+            irisPath.stopFollowingPath();
             var hr = new XMLHttpRequest(),
                 wall1,
                 wall2,
@@ -281,32 +293,23 @@ function playState() {
             hr.setRequestHeader("Content-type", "application/x-www-form-urlencoded");
             hr.send(dataToSend);
             //TODO have players put their name
-            FM.game.switchState(endState())
+            FM.game.switchState(endState(false))
         }
-        if (dead) {
-            //irisRenderer.setWidth(irisRenderer.getWidth() * 1.5);
-        }
-		if (player.physic.overlapsWithObject(eyePhysic)) {
-			finalGoalReached = true;
-		}
-        /*if (wallType.overlapsWithObject(visibilityTestObject)) {
-            console.log("COLLISION!!");
-            visibilityTestObjectPath.stopFollowingPath();
-            checkingVisibility = false;
-            visibilityTestObjectSpatial.position.reset(irisSpatial.position.x - 5, irisSpatial.position.y - 5);
+        //Check if the player won
+        if (!finalGoalReached && !dead && player.physic.overlapsWithObject(eyePhysic)) {
+            console.log("WON");
+            finalGoalReached = true;
+            iris.kill();
+            iris.hide();
+            player.kill();
+            player.hide();
             invisible = true;
-        }*/
-        if (finalGoalReached) {
-			irisSpatial.position.reset(eyeSpatial.position.x + eyePhysic.offset.x + eyePhysic.radius - irisRenderer.getWidth() / 2, eyeSpatial.position.y + eyePhysic.offset.y + eyePhysic.radius - irisRenderer.getHeight() / 2);
-			irisRenderer.setWidth(irisRenderer.getWidth() * 0.99);
-			irisRenderer.setHeight(irisRenderer.getHeight() * 0.99);
-			invisible = true;
-			
-			var hr = new XMLHttpRequest();
+
+            var hr = new XMLHttpRequest();
             hr.open("POST", "src/web/reset.php", true);
             hr.setRequestHeader("Content-type", "application/x-www-form-urlencoded");
             hr.send(dataToSend);
-			FM.game.switchState(menuState());
+            FM.game.switchState(endState(true));
         }
         //Map cursor and modify player's angle
         cursorSpatial.position.x = FM.game.getMouseX();
